feat: helper function
